Use exact versions in Gemfile
Instead of leaving the default version range in Gemfile, used exact
versions for the gems in order to have a more reproducible build. Also
removed the Ruby version as indicated by section 1.2.1 of the tutorial
and removed the 'tzinfo-data' gem which is a patch for Windows. Since we
are not on Windows, we have removed it.