Set the root route (Complete 1.2.4)
The default root route in Rails is "Rails::WelcomeController#index".
However, we have defined a custom controller action (named "hello") in
the Application Controller, and we want that custom controller action to
be rendered whenever the root route is requested. Hence, we have set the
root route to be the "hello" action of the "application" controller in
the 'config/routes.rb' file by adding:

    root 'application#hello'