Update the Gemfile.lock to contain PostgreSQL
Since we specified in Gemfile that PostgreSQL should only be installed
in the production, in fact, there isn't anything new to install in
development. However, we still need to update the Gemfile.lock in order
to have a stable and reproducible system. Hence, in order to update the
Gemfile.lock to contain the PostgreSQL gem, we ran:

    bundle install --without production

This command makes `bundle` to install every gem that is listed in the
Gemfile, except the gems that is listed in the group named "production"
in the Gemfile. Since the only change that we made in the Gemfile was to
add the "production" group, and we didn't add anything else, there
weren't any gems to install. However, Bundler still fetched the gem
version information for each and every gem in the Gemfile, and it has
updated the Gemfile.lock with the versions of every single gem that is
listed in the Gemfile.